# docs/openapi.json
{
  "components": {
    "schemas": {
      "AffiliationCreate": {
        "properties": {
          "department_id": {
            "description": "Department Identifier or Code",
            "maxLength": 50,
            "minLength": 2,
            "title": "Department Id",
            "type": "string"
          },
          "faculty_id": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional Faculty Identifier or Code",
            "title": "Faculty Id"
          },
          "user_id": {
            "description": "Identity Service user id or university id (stored as the canonical user id)",
            "maxLength": 50,
            "minLength": 2,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "department_id"
        ],
        "title": "AffiliationCreate",
        "type": "object"
      },
      "AffiliationListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AffiliationResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "AffiliationListResponse",
        "type": "object"
      },
...
      "AffiliationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "department_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Code"
          },
          "department_id": {
            "title": "Department Id",
            "type": "string"
          },
          "department_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Name"
          },
          "faculty_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Code"
          },
          "faculty_id": {
            "title": "Faculty Id",
            "type": "string"
          },
          "faculty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "department_id",
          "faculty_id",
          "created_at"
        ],
        "title": "AffiliationResponse",
        "type": "object"
      },
      "AffiliationSingleResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AffiliationResponse"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "data"
        ],
        "title": "AffiliationSingleResponse",
        "type": "object"
      },
      "AffiliationUpdate": {
        "properties": {
          "department_id": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Updated Department Identifier or Code",
            "title": "Department Id"
          },
          "faculty_id": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional Updated Faculty Identifier or Code",
            "title": "Faculty Id"
          }
        },
        "title": "AffiliationUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT issued by the Identity Service",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Directory microservice handling Faculties, Departments, Service Units, User Affiliations and Service Responsibilities. Versioned API under /api/v1.",
    "title": "University Directory Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/affiliations": {
      "get": {
        "description": "Retrieve a list of organizational affiliations with optional filters and pagination.",
        "operationId": "list_affiliations_api_v1_affiliations_get",
        "parameters": [
          {
            "description": "Number of items to skip",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Max items to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max items to return",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Filter by department identifier",
            "in": "query",
            "name": "department_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by department identifier",
              "title": "Department Id"
            }
          },
          {
            "description": "Filter by faculty identifier",
            "in": "query",
            "name": "faculty_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by faculty identifier",
              "title": "Faculty Id"
            }
          },
          {
            "description": "Filter by user identifier",
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by user identifier",
              "title": "User Id"
            }
          },
          {
            "description": "Case-insensitive search in user id and department/faculty code and name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search in user id and department/faculty code and name",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Affiliations",
        "tags": [
          "Affiliations"
        ]
      },
      "post": {
        "description": "Create an organizational affiliation linking a user to a department and faculty. The user must exist and be ACTIVE in the Identity Service.",
        "operationId": "create_affiliation_api_v1_affiliations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AffiliationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliationSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create User Affiliation",
        "tags": [
          "Affiliations"
        ]
      }
    },
    "/api/v1/affiliations/users/{user_id}": {
      "get": {
        "description": "Retrieve the current organizational affiliation for a specific user.",
        "operationId": "get_user_affiliation_api_v1_affiliations_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliationSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get User Organizational Affiliation",
        "tags": [
          "Affiliations"
        ]
      }
    },
    "/api/v1/affiliations/{affiliation_id}": {
      "delete": {
        "description": "Permanently delete a user organizational affiliation record.",
        "operationId": "delete_affiliation_api_v1_affiliations__affiliation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "affiliation_id",
            "required": true,
            "schema": {
              "title": "Affiliation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Affiliation",
        "tags": [
          "Affiliations"
        ]
      },
      "get": {
        "description": "Retrieve details for a specific affiliation by its unique identifier.",
        "operationId": "get_affiliation_api_v1_affiliations__affiliation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "affiliation_id",
            "required": true,
            "schema": {
              "title": "Affiliation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliationSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Affiliation by ID",
        "tags": [
          "Affiliations"
        ]
      },
      "put": {
        "description": "Update or transfer a user affiliation to another department/faculty. The affiliated user must still be ACTIVE in the Identity Service.",
        "operationId": "update_affiliation_api_v1_affiliations__affiliation_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "affiliation_id",
            "required": true,
            "schema": {
              "title": "Affiliation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AffiliationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliationSingleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Affiliation",
        "tags": [
          "Affiliations"
        ]
      }
    },
...
  }
}